+++ packages/terra-action-footer/CHANGELOG.md
@@ -4,6 +4,24 @@ ChangeLog
 Unreleased
 ----------
 
+2.32.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.31.0 - (January 17, 2020)
+------------------
+### Changes
+* Update docs to use mdx syntax
+
+### Removed
+* Removed empty examples from doc examples
+
+2.30.0 - (January 7, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.29.0 - (December 10, 2019)
 ------------------
 ### Changed