+++ packages/terra-alert/CHANGELOG.md
@@ -3,6 +3,23 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update docs to use mdx syntax.
+
+4.19.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.18.0 - (January 17, 2020)
+------------------
+### Changed
+* Updated `uncontrolled responsive-element` to `controlled responsive-element`.
+
+4.17.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
 
 4.16.0 - (December 10, 2019)
 ------------------