+++ packages/terra-badge/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update docs to use MDX syntax
+
+3.29.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.28.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
 
 3.27.0 - (December 3, 2019)
 ------------------