+++ packages/terra-button-group/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update docs to use mdx syntax.
+
+3.30.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.29.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
 
 3.28.0 - (December 10, 2019)
 ------------------