+++ packages/terra-card/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update docs to use mdx syntax.
+
+3.23.0 - (January 7, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.22.0 - (December 3, 2019)
 ------------------