+++ packages/terra-dialog/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.32.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.31.0 - (January 7, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.30.0 - (December 10, 2019)
 ------------------
 ### Changed