+++ packages/terra-form-select/CHANGELOG.md
@@ -3,7 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+
+5.40.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+5.39.0 - (January 7, 2020)
+------------------
 ### Fixed 
+* Fixed page scrolling when embedded within an iframe.
 * Fixed accessibility issue due to invalid `aria-activedescendant` value.
 
 5.38.0 - (December 3, 2019)