+++ packages/terra-heading/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+4.23.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.22.0 - (January 7, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.21.0 - (December 3, 2019)
 ------------------
 ### Added