+++ packages/terra-markdown/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.37.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
+
 2.36.0 - (December 3, 2019)
 ------------------
 ### Added