+++ packages/terra-overlay/CHANGELOG.md
@@ -3,8 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+
+3.39.0 - (January 28, 2020)
+------------------
 ### Changed
-* Update screenshot
+* Minor dependency version bump
+
+3.38.0 - (January 7, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.37.0 - (December 10, 2019)
 ------------------