+++ packages/terra-search-field/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added onInput prop.
+
+3.37.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.36.0 - (January 7, 2020)
+------------------
+### Changed
+* Updated wdio screenshots
 
 3.35.0 - (December 10, 2019)
 ------------------