+++ packages/terra-spacer/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+3.28.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.27.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
+
 3.26.0 - (December 10, 2019)
 ------------------
 ### Changed