+++ packages/terra-status-view/CHANGELOG.md
@@ -3,6 +3,17 @@ ChangeLog
 
 Unreleased
 ----------
+
+4.15.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.14.0 - (January 7, 2020)
+------------------
+### Removed
+* Removed yarn reference in docs
+
 ### Changed
 * replaced css grid with flex-box